requirement.txt
Django
djangorestframework
pymysql
mysqlclient
PyJWT